fix(header-footer): update shadowed header and footer for gatsby 2
