+++ src/gatsby-theme-carbon/components/Header/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Header from 'gatsby-theme-carbon/src/components/Header';
 
-const CustomHeader = props => (
+const CustomHeader = (props) => (
   <Header {...props}>
     <span>Carbon for IBM.com</span>
   </Header>
